requirements.txt
numpy
pillow_heif
opencv-python
matplotlib
numpy
Pillow
torch
torchvision
tqdm
pyyaml
requests
pandas
seaborn
uvicorn
datetime
pytest
fastapi
python-multipart
httpx
boto3
pytest-cov